// client/src/autobindings/types.ts
export const ActiveMonsterAttack = __t.object("ActiveMonsterAttack", {
  activeMonsterAttackId: __t.u64(),
  scheduledAt: __t.scheduleAt(),
  get position() {
    return DbVector2;
  },
  get direction() {
    return DbVector2;
  },
  get monsterAttackType() {
    return MonsterAttackType;
  },
  piercing: __t.bool(),
  damage: __t.u32(),
  radius: __t.f32(),
  speed: __t.f32(),
  parameterU: __t.u32(),
  parameterF: __t.f32(),
  ticksElapsed: __t.u32(),
  fromShinyMonster: __t.bool(),
});
export type ActiveMonsterAttack = __Infer<typeof ActiveMonsterAttack>;
// ...
export const DbVector2 = __t.object("DbVector2", {
  x: __t.f32(),
  y: __t.f32(),
});
export type DbVector2 = __Infer<typeof DbVector2>;
// ...
export const MonsterAttackType = __t.enum("MonsterAttackType", {
  ImpBolt: __t.unit(),
  EnderBolt: __t.unit(),
  EnderScytheSpawn: __t.unit(),
  EnderScythe: __t.unit(),
  ChaosBall: __t.unit(),
  VoidZone: __t.unit(),
  AgnaFlamethrowerJet: __t.unit(),
  AgnaOrbSpawn: __t.unit(),
  AgnaFireOrb: __t.unit(),
  AgnaCandleBolt: __t.unit(),
  AgnaPhase2FlameJet: __t.unit(),
  AgnaGroundFlame: __t.unit(),
  SimonChemicalBolt: __t.unit(),
  SimonToxicZone: __t.unit(),
  SimonToxicSpray: __t.unit(),
});
export type MonsterAttackType = __Infer<typeof MonsterAttackType>;